+++ lib/src/features/collection/data/local/local_collection_repo.dart
@@ -199,6 +199,7 @@ class LocalCollectionRepo {
     }
 
     _sharedPref.remove(DBKeys.SAVED_COLLECTION_SECTIONS);
+    _localHomeRepo.clearAllLocalCachedPreviews();
   }
 
   //